renamed interaction to _interaction
