+++ pyfar/plot/_interaction.py
